refactor(eql-bindings): split DomainType trait + all() out of mod.rs
Move the trait/consts/PhantomData impl into hand-written domain_type.rs and the
all() inventory into inventory.rs (re-exported at crate::v3, public paths
unchanged). Relocate the non-catalog-derivable float NaN and bool storage-only
caveats into the hand-written mod.rs doc so the generated per-family modules
don't drop them at cutover. mod.rs stays hand-written. types:check green.
